requirements.txt
# =================================================================
#  Core Deep Learning Framework
# =================================================================
# torch, torchvision, torchaudio는 사용자의 CUDA 버전에 맞게
# 별도로 설치하는 것을 권장합니다. (아래 버전은 cu118 기준)
torch==2.4.1+cu118
torchvision==0.19.1+cu118
torchaudio==2.4.1+cu118
numpy==1.24.1

# =================================================================
#  Model Evaluation & Metrics
# =================================================================
# IS, FID, KID 등 정량적 평가 지표 계산을 위한 라이브러리
torchmetrics==1.5.2
scipy==1.10.1  # Inception Score 계산에 torchmetrics가 의존함
torch-fidelity==0.3.0 # InceptionScore, FID, KID 계산에 필요

# =================================================================
#  Monitoring & Visualization
# =================================================================
# 학습 과정 모니터링 (손실 그래프 등)
tensorboard==2.14.0
# 유틸리티 스크립트에서 시각화에 사용될 수 있음
matplotlib==3.7.5

# =================================================================
#  Utilities
# =================================================================
# 터미널에 진행률 바(Progress Bar) 표시
tqdm==4.67.1
# 이미지 로딩 및 처리를 위한 라이브러리
Pillow==10.2.0